requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=h93e1e8c_2
aiofiles=23.2.1=pypi_0
aiohttp=3.9.5=py310h2372a71_0
aiosignal=1.3.1=pyhd8ed1ab_0
altair=5.3.0=pypi_0
annotated-types=0.7.0=pyhd8ed1ab_0
anyio=4.3.0=pyhd8ed1ab_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
backoff=2.2.1=pypi_0
bcrypt=4.1.3=pypi_0
beautifulsoup4=4.12.3=pypi_0
brotli-python=1.1.0=py310hc6cd4ac_1
build=1.2.1=pypi_0
bzip2=1.0.8=hd590300_5
c-ares=1.28.1=hd590300_0
ca-certificates=2024.6.2=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.3=pyhd8ed1ab_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py310h2fee648_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
chevron=0.14.0=pyhd3deb0d_1
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.5.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cryptography=42.0.8=py310hb1bd9d3_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.6=pyhd8ed1ab_0
debugpy=1.8.1=py310hc6cd4ac_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.3=pypi_0
ffmpy=0.3.2=pypi_0
fonttools=4.51.0=pypi_0
frozendict=2.4.4=py310hc51659f_0
frozenlist=1.4.1=py310h2372a71_0
genanki=0.13.1=pyhd8ed1ab_0
google-ai-generativelanguage=0.6.3=pypi_0
google-api-core=2.19.0=pyhd8ed1ab_0
google-api-python-client=2.129.0=pypi_0
google-auth=2.29.0=pyhca7485f_0
google-auth-httplib2=0.2.0=pyhd8ed1ab_0
google-generativeai=0.5.3=pypi_0
googleapis-common-protos=1.63.0=pypi_0
gradio=4.31.3=pypi_0
gradio-client=0.16.3=pypi_0
greenlet=3.0.3=py310hc6cd4ac_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.64.1=pypi_0
grpcio-status=1.62.2=pypi_0
httpcore=1.0.5=pypi_0
httplib2=0.22.0=pyhd8ed1ab_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jsonpatch=1.33=pyhd8ed1ab_0
jsonpointer=2.4=py310hff52083_3
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.2=h659d440_0
kubernetes=29.0.0=pypi_0
langchain=0.1.20=pyhd8ed1ab_0
langchain-community=0.0.38=pyhd8ed1ab_0
langchain-core=0.1.52=pyhd8ed1ab_0
langchain-google-genai=1.0.3=pyhd8ed1ab_1
langchain-text-splitters=0.0.2=pyhd8ed1ab_0
langsmith=0.1.71=pyhd8ed1ab_0
ld_impl_linux-64=2.40=hf3520f5_2
libabseil=20211102.0=cxx17_h48a1fff_2
libblas=3.9.0=22_linux64_openblas
libcblas=3.9.0=22_linux64_openblas
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_7
libgfortran-ng=13.2.0=h69a702a_7
libgfortran5=13.2.0=hca663fb_7
libgomp=13.2.0=h77fa898_7
liblapack=3.9.0=22_linux64_openblas
libnsl=2.0.1=hd590300_0
libopenblas=0.3.27=pthreads_h413a1c8_0
libprotobuf=3.20.3=h3eb15da_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=hc0a3c3a_7
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=h4ab18f5_6
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
marshmallow=3.21.2=pyhd8ed1ab_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=py310h2372a71_0
mypy_extensions=1.0.0=pyha770c72_0
ncurses=6.5=h59595ed_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=1.26.4=py310hb13e2d6_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.18.0=pypi_0
openssl=3.3.1=h4ab18f5_0
opentelemetry-api=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.25.0=pypi_0
opentelemetry-instrumentation=0.46b0=pypi_0
opentelemetry-instrumentation-asgi=0.46b0=pypi_0
opentelemetry-instrumentation-fastapi=0.46b0=pypi_0
opentelemetry-proto=1.25.0=pypi_0
opentelemetry-sdk=1.25.0=pypi_0
opentelemetry-semantic-conventions=0.46b0=pypi_0
opentelemetry-util-http=0.46b0=pypi_0
orjson=3.10.3=py310he421c4c_0
overrides=7.7.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
posthog=3.5.0=pypi_0
prompt-toolkit=3.0.46=pyha770c72_0
proto-plus=1.23.0=pyhd8ed1ab_0
protobuf=3.20.3=py310h6a678d5_0
psutil=5.9.8=py310h2372a71_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pyhd8ed1ab_0
pyasn1-modules=0.4.0=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pymupdf=1.24.3=pypi_0
pymupdfb=1.24.3=pypi_0
pyopenssl=24.0.0=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.14=hd12c33a_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.10=4_cp310
pytz=2024.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0.1=py310h2372a71_1
pyzmq=26.0.3=py310h6883aea_0
re2=2022.04.01=h27087fc_0
readline=8.2=h8228510_1
referencing=0.35.1=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pyhd8ed1ab_0
ruff=0.4.4=pypi_0
scipy=1.13.1=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=70.0.0=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.30=py310hc51659f_0
sqlite=3.45.3=h2c6b66d_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
statsmodels=0.14.2=pypi_0
sympy=1.12.1=pypi_0
tenacity=8.3.0=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4=py310h2372a71_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.3=pypi_0
typing-extensions=4.12.1=hd8ed1ab_0
typing_extensions=4.12.1=pyha770c72_0
typing_inspect=0.9.0=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
ujson=5.10.0=pypi_0
uritemplate=4.1.1=pyhd8ed1ab_0
urllib3=2.2.1=pyhd8ed1ab_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.9.4=py310h2372a71_0
zeromq=4.3.5=h75354e8_4
zipp=3.19.2=pypi_0
zlib=1.2.13=h4ab18f5_6
